triptodos/requirements.txt
python-etcd
python-dateutil
